requirements.txt
faiss-cpu
sentence-transformers
transformers
streamlit
pylatexenc